Fix scaling of apng images (#143)
A previous change resulted in some apng images not being of uint8 format, which displayed wrong.

This corrects that, and also allows the images service to take in simple messages

+++ tests/services/test_images_plugins.py
@@ -183,8 +183,13 @@ def test_mrc_to_apng_skip_rescaling(mock_pil, tmp_path):
     # Check the image creation
     assert mock_pil.fromarray.call_count == 2
     mock_pil.fromarray.assert_called_with(mock.ANY, mode="L")
-    assert (mock_pil.fromarray.mock_calls[0][1] == data_3d[0] * 255 / 24).all()
-    assert (mock_pil.fromarray.mock_calls[2][1] == (data_3d[1] - 25) * 255 / 24).all()
+    assert (
+        mock_pil.fromarray.mock_calls[0][1] == (data_3d[0] * 255 / 24).astype("uint8")
+    ).all()
+    assert (
+        mock_pil.fromarray.mock_calls[2][1]
+        == ((data_3d[1] - 25) * 255 / 24).astype("uint8")
+    ).all()
     mock_pil.fromarray().thumbnail.assert_called_with((512, 512))
